Move initialisation of controller into separate file
